weather stations home page work in progress

+++ src/WeatherMonitorCore/WeatherMonitorCore.MqttDataSubscriberService.Interfaces/Models/MeasurementsLog.cs
@@ -2,10 +2,10 @@
 
 public readonly record struct MeasurementsLog(
     DateTime ReceivedAt,
-    float Humidity,
-    float Temperature,
-    float AirPressure,
-    float Altitude,
-    float? PM1_0,
-    float? PM2_5,
-    float? PM10);
+    double Humidity,
+    double Temperature,
+    double AirPressure,
+    double Altitude,
+    double? PM1_0,
+    double? PM2_5,
+    double? PM10);
